Refactor Project Banner Integration to Use Client Adapter
- Replaced ProjectBannerAdapter with ProjectBannerClientAdapter across multiple files for improved data handling.
- Introduced ProjectBannerClientAdapter and ProjectBannerClientAdapterMock to manage project banner data retrieval.
- Updated related imports and types to reflect the new client adapter structure.
- Added JSON data source for project banners to enhance mock data capabilities.
