+++ core/infrastructure/json-storage-client-adapter/contract/api.d.ts
