+++ core/infrastructure/json-storage-client-adapter/data/project-banner-client-data.json
