+++ core/infrastructure/json-storage-client-adapter/mock-adapters/project-banner-client-adapter-mock.ts
@@ -2,7 +2,9 @@ import { ProjectBanner } from "@/core/domain/project-banner/models/project-banne
 import { ProjectBannerStoragePort } from "@/core/domain/project-banner/outputs/project-banner-storage-port";
 import { GetProjectBannersPortResponse } from "@/core/domain/project-banner/project-banner-contract.types";
 
-export class ProjectBannerAdapterMock implements ProjectBannerStoragePort {
+export class ProjectBannerClientAdapterMock implements ProjectBannerStoragePort {
+  constructor() {}
+
   routes = {};
 
   getProjectBanners(): GetProjectBannersPortResponse {
